% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amalgamate.R
\name{amalgamate}
\alias{amalgamate}
\title{amalgamate the neighborhood site}
\usage{
amalgamate(geno, window.len = 500, delim = "_")
}
\arguments{
\item{geno}{genotype vecotor, with only 0,1,2}

\item{window.len}{window length to merge the genotype}

\item{delim}{delimeter to split the position, e.g the delim of
"chr_01" should be "_"}
}
\value{
a amalgamated genotype vector
}
\description{
amalgamate the neighborhood site
}
\examples{
geno <- rep(c(rep(1, 10),0, rep(1,3),2,rep(2,3)), 2)
names(geno) <- c(paste0("chr1_", c(1,10:15, 20, 30:38, 51)),
              paste0("chr2_", c(1,10:15, 20, 30:38, 51)))
geno
amalgamate(geno, 8)

}
\author{
Zhougeng Xu
}
