% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCallGeno.R
\name{batchCallGeno}
\alias{batchCallGeno}
\title{genotyping all sample by chromsome}
\usage{
batchCallGeno(x, CHROM, outdir = ".", window.size = 15,
  low.count = 6, high.count = 24, fix.size = 5, pdf.height = 4,
  pdf.width = 8)
}
\arguments{
\item{x}{GT matrix}

\item{CHROM}{chromosome vector}

\item{outdir}{outdir}

\item{window.size}{parameter of callWindowGeno}

\item{low.count}{parameter of callWindowGeno}

\item{high.count}{parameter of callWindowGeno}

\item{fix.size}{parameter of fixGenoError}

\item{pdf.height}{pdf width}

\item{pdf.width}{pdf width}
}
\description{
genotyping all sample by chromsome
}
\author{
Zhou-Geng xu
}
