% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGtFromAd.R
\name{calcFreqFromAd}
\alias{calcFreqFromAd}
\title{Calculate the frequcy of AD}
\usage{
calcFreqFromAd(x, min.depth = 10, max.depth = 200)
}
\arguments{
\item{x}{AD matrix}

\item{min.depth}{minimum depth to infer the genotype, if
depth lower than it, it will be conside as NA}

\item{max.depth}{maximum depth to infer the genotype, if
depth larger than it, it will be conside as NA}
}
\value{
matrix contains. The alt/(alt+ref) will be caculted
from the allele depth
}
\description{
Calculate the frequcy of AD
}
\examples{
AD <- matrix(data = c("30,1","1,30","0,0","15,15"), nrow = 2)
row.names(AD) <- c("chr1_1","chr1_100")
colnames(AD) <- c("A","B")
freq <- calcFreqFromAd(AD)

}
\author{
Zhou-geng Xu
}
