% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{calcMissRatio}
\alias{calcMissRatio}
\title{Calculate the missing ratio of each sample on all site}
\usage{
calcMissRatio(x)
}
\arguments{
\item{x}{a binmapr object}
}
\value{
x a binmapr object with ind.miss and mar.miss
}
\description{
Calculate the missing ratio of each sample on all site
}
\author{
Zhou-geng Xu
}
