% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callWindowGeno.R
\name{callWindowGeno}
\alias{callWindowGeno}
\title{call window genotype}
\usage{
callWindowGeno(x, window.size = 15, low = 6, high = 24)
}
\arguments{
\item{x}{a vector object, storing genotype information}

\item{window.size}{default is 15}

\item{low}{default is 6}

\item{high}{default is 24}
}
\description{
call window genotype
}
\author{
Zhougeng Xu
}
