% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export2asmap}
\alias{export2asmap}
\title{Generate input for mstmap from callWindowGeno result}
\usage{
export2asmap(geno, pop.type = "DH")
}
\arguments{
\item{geno}{gentype matrix or list from callWindowGeno}

\item{pop.type}{Character string specifying the population
type of the data frame object. Accepted values are "DH"
(doubled haploid), "BC" (backcross), "RILn" (non-advanced
RIL population with n generations of selfing) and "ARIL"
(advanced RIL) (see Details). Default is "DH".}
}
\value{
data.frame
}
\description{
Generate input for mstmap from callWindowGeno result
}
\examples{
data(geno)
GT <- as.matrix(geno[1:5,5:10])
rownames(GT) <- paste0(geno$CHR[1:5], "_", geno$POS[1:5])
df <- export2asmap(GT)

}
\author{
Zhougeng Xu
}
