% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export2cross}
\alias{export2cross}
\title{Generate cross object for R/qtl}
\usage{
export2cross(
  x,
  BC.gen = 0,
  F.gen = 0,
  alleles = c("A", "B"),
  parent1 = NULL,
  parent2 = NULL
)
}
\arguments{
\item{x}{}

\item{BC.gen}{Used only for cross type "bcsft"}

\item{F.gen}{Used only for cross type "bcsft"}
}
\value{
data.frame
}
\description{
Generate cross object for R/qtl
}
\examples{
data(geno)
GT <- as.matrix(geno[1:5,5:10])
rownames(GT) <- paste0(geno$CHR[1:5], "_", geno$POS[1:5])
df <- export2asmap(GT)

}
\author{
Zhougeng Xu
}
