% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixGenoError.R
\name{fixGenoError}
\alias{fixGenoError}
\title{fix potential genotype error}
\usage{
fixGenoError(geno, fix.size = 10)
}
\arguments{
\item{geno}{a vector object, storing genotyping information}

\item{fix.size}{the size of short genotpe error}
}
\description{
fix potential genotype error
}
\author{
Zhou-geng Xu, Guang-wei Li
}
