% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geno}
\alias{geno}
\title{Rice QTL genotype data}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(geno)
}
\description{
Data from "Three representative inter and intra-subspecific crosses
 reveal the genetic architecture of reproductive isolation in rice."
}
\examples{
data(geno)
}
\references{
Li, G. et al. (2017) The Plant Journal 92, 349–362.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28805257}{PubMed})
}
\keyword{datasets}
