% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGeno}
\alias{plotGeno}
\title{genotype plotting}
\usage{
plotGeno(geno, pos.start = 6, xlab = "position", ylab = NULL,
  title = NULL)
}
\arguments{
\item{geno}{genotype vector}

\item{pos.start}{position start index,
for exmaple, the pos.start of chr1_1234 is 6}

\item{xlab}{x lab}

\item{ylab}{y lab}

\item{title}{title}
}
\description{
genotype plotting
}
\author{
Zhougeng Xu
}
