% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGenoHeatmap}
\alias{plotGenoHeatmap}
\title{genotype heatmap}
\usage{
plotGenoHeatmap(
  obj,
  samples = NULL,
  markers = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL
)
}
\description{
genotype heatmap
}
