% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMarker}
\alias{plotMarker}
\title{marker distribution in scaffold/contigs}
\usage{
plotMarker(obj, min_size = 1e+05, min_num = 10, step_size = 1e+06)
}
\arguments{
\item{obj}{binmapr object}

\item{ctg}{a vector contain contig name}

\item{ort}{a vector contain contig orientation}

\item{delim}{delimiter of marker name}
}
\description{
marker distribution in scaffold/contigs
}
