% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotQtl}
\alias{plotQtl}
\title{QTL mapping plot}
\usage{
plotQtl(pos, pvalue, ylab = "LOD", chr.name = "", ymax = 10,
  threshold = 2, ...)
}
\arguments{
\item{pos}{position of each p value}

\item{pvalue}{p value}

\item{ylab}{y lab}

\item{chr.name}{chromosome name}

\item{ymax}{y max}

\item{threshold}{QTL threshold}

\item{...}{other parameter pass to base::plot}
}
\description{
QTL mapping plot
}
\author{
Zhougeng Xu
}
