% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRfHeatmap}
\alias{plotRfHeatmap}
\title{Recombination fractions heatmap}
\usage{
plotRfHeatmap(obj, ctg, ori)
}
\arguments{
\item{obj}{binmapr object}

\item{ctg}{a vector contain contig name}

\item{ort}{a vector contain contig orientation}
}
\description{
Recombination fractions heatmap
}
