% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{[.binmapr}
\alias{[.binmapr}
\title{Subset method for binmapr objects}
\usage{
\method{[}{binmapr}(x, i, j, by, ...)
}
\arguments{
\item{x}{A binmapr object.}

\item{i}{A numeric vector or logical vector to subset rows.}

\item{j}{A numeric vector or logical vector to subset columns.}

\item{by}{A character string to specify the subset method. Can be either "ind" (individual) or "mar" (marker).}

\item{...}{Other arguments to be passed to the function.}
}
\value{
A subsetted binmapr object.
}
\description{
This function subsets binmapr objects either by individual or by marker.
}
\examples{
# Create a binmapr object
bm <- binmapr(...)
# Subset the binmapr object by individual
bm1 <- bm[1:5, , by="ind"]
# Subset the binmapr object by marker
bm2 <- bm[ , 1:5, by="mar"]
}
