% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{subset.binmapr}
\alias{subset.binmapr}
\title{subset the binmaprt data}
\usage{
\method{subset}{binmapr}(x, by, ...)
}
\arguments{
\item{x}{a binmapr object}

\item{by}{subset data by "ind"(sample) or "mar"(marker)}
}
\description{
subset the binmaprt data
}
