% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPHASE.R
\name{write.fastphase}
\alias{write.fastphase}
\title{write fastPHASE format}
\usage{
write.fastphase(obj, file = "")
}
\arguments{
\item{obj}{a binmapr object}

\item{file}{output file name}
}
\description{
write fastPHASE format
}
