% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCallGeno.R
\name{batchCallGeno}
\alias{batchCallGeno}
\title{Genotyping all sample by chromsome}
\usage{
batchCallGeno(
  x,
  CHROM,
  min.marker.nums = 10,
  collapsed = FALSE,
  outdir = ".",
  window.len = 1000,
  delim = "_",
  window.size = 15,
  low.count = 6,
  high.count = 24,
  fix.size = 5,
  pos.start = 6,
  plot.geno = TRUE,
  pdf.height = 4,
  pdf.width = 8
)
}
\arguments{
\item{x}{GT matrix}

\item{CHROM}{chromosome vector}

\item{min.marker.nums}{the minimum marker number on the chromosome}

\item{collapsed}{merge the neighbor SNPs into single site}

\item{outdir}{outdir}

\item{window.len}{window length to merge the genotype}

\item{delim}{delimeter to split the position, e.g the delim of
"chr_01" should be "_"}

\item{window.size}{default is 15}

\item{fix.size}{define the neighbor size to fix the error}

\item{pos.start}{position start index,
for exmaple, the pos.start of chr1_1234 is 6, chr01_1234 is 7}

\item{plot.geno}{whether to plot the genotype or not}

\item{pdf.height}{pdf width}

\item{pdf.width}{pdf width}
}
\value{
A list object including genotype of each chromosome. It will
also create PDF and CSV of these genotype.
}
\description{
The genotype of each sample will be called according to
their allele depth, and then the potential error will be fixed.
}
\examples{
data(geno)
GT <- as.matrix(geno[1:100,5:6])
row.names(GT) <- paste0(geno$CHR[1:100], "_", geno$POS[1:100])
gt <- batchCallGeno(GT, "chr01",pos.start = 7)

}
\author{
Zhougeng xu
}
