% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMstmapInput.R
\name{buildMstmapInput}
\alias{buildMstmapInput}
\title{Generate input for mstmap from callWindowGeno result}
\usage{
buildMstmapInput(geno)
}
\arguments{
\item{geno}{gentype matrix}
}
\value{
data.frame
}
\description{
Generate input for mstmap from callWindowGeno result
}
\examples{
data(geno)
GT <- as.matrix(geno[1:5,5:10])
rownames(GT) <- paste0(geno$CHR, "_", geno$POS)[1:5]
df <- buildMstmapInput(GT)

}
\author{
Zhougeng Xu
}
