% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGtFromAd.R
\name{calcFreqFromAd}
\alias{calcFreqFromAd}
\title{Calculate the frequcy of AD}
\usage{
calcFreqFromAd(x, min.depth = 10)
}
\arguments{
\item{x}{AD matrix}

\item{min.depth}{minimum depth to infer the genotype, if
depth lower than it, it will be conside as NA}
}
\description{
Calculate the frequcy of AD
}
\author{
Zhou-geng Xu
}
