% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGtFromAd.R
\name{callGtFromAd}
\alias{callGtFromAd}
\title{call genotype from Allele depth with updog}
\usage{
callGtFromAd(x, ploidy = 2, model = "norm", n.cpus = 1)
}
\arguments{
\item{x}{a binmapr object}

\item{ploidy}{The ploidy of the species. Assumed to be the same for each individual.}

\item{model}{What form should the prior (genotype distribution) take?
More detail in help document of updog::multidog}

\item{n.cpus}{The number of computing cores to use}
}
\value{
binmapr object, containing  genotype slot
}
\description{
call genotype from Allele depth with updog
}
\details{
call genotype with updog
}
\examples{
AD <- matrix(data = c("30,1","1,30","0,0","15,15"), nrow = 2)
row.names(AD) <- c("chr1_1","chr1_100")
colnames(AD) <- c("A","B")
geno <- callGtFromAd(AD)

}
\author{
Zhougeng Xu
}
