% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGtFromAd.R
\name{callGtFromAd}
\alias{callGtFromAd}
\title{call genotype from AD}
\usage{
callGtFromAd(x, min.depth = 10, max.depth = 200, low = 0.2,
  high = 0.8)
}
\arguments{
\item{x}{AD matrix}

\item{min.depth}{minimum depth to infer the genotype, if
depth lower than it, it will be conside as NA}

\item{max.depth}{maximum depth to infer the genotype, if
depth larger than it, it will be conside as NA}

\item{low}{threshold to infer one parent, encoded as 0}

\item{high}{threshold to infer another parent, encoded as 2}
}
\value{
matrix. The allele depth will convet to genotype
}
\description{
call genotype from AD
}
\details{
The original GT in VCF may be wrong, if the variant is sequencing error.
To avoid this kind of error, we used the ALT/(REF + ALT) to infer
the genotype. If the ratio <= 0.2, it will be encoded as 0. If
the ratio >= 0.8, it will be encoded as 2. Otherwise, it will be
encoded as 1.
}
\examples{
AD <- matrix(data = c("30,1","1,30","0,0","15,15"), nrow = 2)
row.names(AD) <- c("chr1_1","chr1_100")
colnames(AD) <- c("A","B")
geno <- callGtFromAd(AD)

}
\author{
Zhougeng Xu
}
