% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGtFromAd.R
\name{callGtFromAd}
\alias{callGtFromAd}
\title{call genotype from AD}
\usage{
callGtFromAd(x, min.depth = 10, low = 0.2, high = 0.8)
}
\arguments{
\item{x}{AD matrix}

\item{min.depth}{minimum depth to infer the genotype, if
depth lower than it, it will be conside as NA}

\item{low}{threshold to infer one parent, 0}

\item{high}{threshold to infer another parent, 2}
}
\description{
call genotype from AD
}
\author{
Zhou-geng Xu
}
