% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callWindowGeno.R
\name{callWindowGeno}
\alias{callWindowGeno}
\title{Call genotype by fix-size window}
\usage{
callWindowGeno(binm, window.size = 15, low = 6, high = 24)
}
\arguments{
\item{binm, }{a binmapr object}

\item{window.size}{default is 15}

\item{low}{default is 6}

\item{high}{default is 24}
}
\value{
genotype matrix list
}
\description{
Call genotype by fix-size window
}
\author{
Zhougeng Xu
}
