% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixGenoError.R
\name{fixGenoError}
\alias{fixGenoError}
\title{Fix potential genotype error}
\usage{
fixGenoError(geno, fix.size = 10)
}
\arguments{
\item{geno}{a vector object, storing genotyping information}

\item{fix.size}{the size of short genotpe error}
}
\value{
vector contains error-fixed genotype
}
\description{
Fix potential genotype error
}
\examples{
genos <- c(1,1,1,1,1,0,1,1,1,1,1,0)
fixGenoError(genos, fix.size = 2)

}
\author{
Zhougeng Xu, Guangwei Li
}
