% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdFromVcf.R
\name{getAdFromVcf}
\alias{getAdFromVcf}
\title{get the AD from vcf}
\usage{
getAdFromVcf(file, keep = NULL, chromosome = NULL)
}
\arguments{
\item{file}{vcf file path}

\item{keep}{a vector store high confidence site, format should be "chr_1"}

\item{chromosome}{vector, which chromosome to use
if it is NULL, all chromosome will be include in analysis}
}
\description{
get the AD from vcf
}
\author{
Zhou-geng Xu
}
