% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotQtl}
\alias{plotQtl}
\title{QTL mapping plot}
\usage{
plotQtl(pos, pvalue, chr.name = NULL, threshold = 2, ...)
}
\arguments{
\item{pos}{position of each p value}

\item{pvalue}{p value}

\item{chr.name}{chromosome name}

\item{threshold}{QTL threshold}

\item{xlab}{x lab}

\item{ylab}{y lab}
}
\description{
QTL mapping plot
}
\author{
Zhou-geng Xu
}
