% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2binmap.R
\name{vcf2binmap}
\alias{vcf2binmap}
\title{convet vcf to binmap automatically}
\usage{
vcf2binmap(vcf, outdir = ".", keep = NULL, chromosome = NULL,
  min.depth = 10, low.ratio = 0.2, high.ratio = 0.8,
  miss.ratio = 0.05, window.size = 15, low.count = 6,
  high.count = 24, fix.size = 5, ...)
}
\arguments{
\item{vcf}{vcf file}

\item{keep}{high confidence snp site}

\item{chromosome}{vector, which chromosome to use
if it is NULL, all chromosome will be include in analysis}

\item{min.depth}{parameter of callGtFromAd}

\item{low.ratio}{parameter of callGtFromAd}

\item{high.ratio}{parameter of callGtFromAd}

\item{miss.ratio}{maximum missing ratio}

\item{window.size}{parameter of callWindowGeno}

\item{low.count}{parameter of callWindowGeno}

\item{high.count}{parameter of callWindowGeno}
}
\description{
convet vcf to binmap automatically
}
\author{
Zhou-geng xu
}
